Moving user creation on a provider
