+++ src/Addons/User/AddonsUserInterface.php
@@ -17,10 +17,11 @@
  * @copyright Since 2007 PrestaShop SA and Contributors
  * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License version 3.0
  */
+declare(strict_types=1);
 
-namespace PrestaShop\Module\Mbo\Addons;
+namespace PrestaShop\Module\Mbo\Addons\User;
 
-interface UserInterface
+interface AddonsUserInterface
 {
     /**
      * @return bool
